gcp/requirements.txt
tensorflow==2.12.0
google-cloud-storage==1.16.1
Pillow==9.5.0